requirements.txt
openai>=1.0.0
boto3>=1.28.0
tqdm>=4.67.1
python-dateutil>=2.8.2
pydub>=0.25.1
